Guard OpenMP user-defined reductions for MSVC and add CMake CI
MSVC's OpenMP support is stuck at OpenMP 2.0, which does not allow
user-defined reductions or unsigned integer loop variables. The
conda-forge feedstock has been carrying a patch to work around this
since the conda-forge MSVC build enables OpenMP via the standalone
CMake target (see r-barnes/richdem#13 follow-up).

Upstream the patch so the feedstock can drop it:
* depression_hierarchy.hpp: gate the std::vector reduction pragmas
  behind RICHDEM_HAS_OMP_VECTOR_REDUCTION, which is 1 only when
  _OPENMP >= 201307 (OpenMP 4.0+) or we are not on MSVC.
* depression_hierarchy.hpp / rd_depression_hierarchy.cpp /
  rd_fill_spill_merge.cpp: switch the three OpenMP `for` loops with
  `unsigned int` indices to signed int64_t (MSVC OpenMP 2.0 requires
  signed loop variables).

Also add a `cmake-msvc-openmp` job to the CI matrix that configures
with `-DUSE_GDAL=OFF -DWITH_TESTS=ON`, builds the `richdem` shared lib
and `richdem_unittests` target with MSVC + /openmp, and runs the unit
tests (skipping the four GDAL-dependent cases). The build itself is
what catches MSVC + OpenMP regressions before they hit the feedstock.

+++ apps/rd_depression_hierarchy.cpp
@@ -36,9 +36,10 @@ int main(int argc, char** argv) {
   rd::Array2D<rd::flowdir_t> flowdirs(topo.width(), topo.height(), rd::NO_FLOW);  // No cells flow anywhere
 
 // Label the ocean cells. This is a precondition for using
-//`GetDepressionHierarchy()`.
+//`GetDepressionHierarchy()`. MSVC's OpenMP 2.0 only allows signed integer
+// loop variables, so use int64_t rather than unsigned int.
 #pragma omp parallel for
-  for (unsigned int i = 0; i < label.size(); i++) {
+  for (int64_t i = 0; i < static_cast<int64_t>(label.size()); i++) {
     // Ocean Level is assumed to be lower than any other cells (even Death Valley)
     if (topo.isNoData(i) || topo(i) == ocean_level) {
       label(i) = dh::OCEAN;
